+++ packages/camera/camera_platform_interface/lib/src/types/types.dart
@@ -7,3 +7,4 @@ export 'resolution_preset.dart';
 export 'camera_exception.dart';
 export 'flash_mode.dart';
 export 'image_format_group.dart';
+export 'exposure_mode.dart';